tests should be together *grumpy face*
